pyproject.toml
[tool.poetry]
name = "skim"
version = "0.1.0"
description = "A simple N Grams prediction model"
authors = ["Tanay PrabhuDesai <tanayseven@gmail.com>"]
license = "GPLv3"

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.18.1"
click = "^7.0"
pyzmq = "^18.1.1"
jsonschema = "^3.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.3.5"
pytest-cov = "^2.8.1"
coveralls = "^1.11.1"
pytest-watch = "^4.2.0"
rope = "^0.16.0"
autopep8 = "^1.5"
black = "^19.10b0"
mypy = "^0.770"
waiting = "^1.4.1"
pre-commit = "^2.3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
